# Mimir query engine: add support for binary operations between instant vectors and scalars