--- a/pkg/streamingpromql/benchmarks/benchmarks.go
+++ b/pkg/streamingpromql/benchmarks/benchmarks.go
@@ -148,6 +148,9 @@ func TestCases(metricSizes []int) []BenchCase {
 		{
 			Expr: "nh_X / a_X",
 		},
+		{
+			Expr: "2 * a_X",
+		},
 		// Test the case where one side of a binary operation has many more series than the other.
 		{
 			Expr: `a_100{l=~"[13579]."} - b_100`,