… CHANGELOG.md
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/operators/vector_vector_binary_operation_test.go
… pkg/streamingpromql/testdata/upstream/aggregators.test
… pkg/streamingpromql/testdata/upstream/functions.test
… pkg/streamingpromql/testdata/upstream/operators.test
… pkg/streamingpromql/testdata/upstream/trig_functions.test